chore(container): refactors the container utility for better code structure.
