+++ sdk/google.golang.org/grpc/server.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	"github.com/traceableai/goagent/sdk"
-	"github.com/traceableai/goagent/sdk/internal"
+	"github.com/traceableai/goagent/sdk/internal/container"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/stats"
 )
@@ -20,7 +20,7 @@ func WrapUnaryServerInterceptor(
 	defaultAttributes := map[string]string{
 		"rpc.system": "grpc",
 	}
-	if containerID, err := internal.GetContainerID(); err == nil {
+	if containerID, err := container.GetID(); err == nil {
 		defaultAttributes["container_id"] = containerID
 	}
 
@@ -175,7 +175,7 @@ func WrapStatsHandler(delegate stats.Handler, spanFromContext sdk.SpanFromContex
 	defaultAttributes := map[string]string{
 		"rpc.system": "grpc",
 	}
-	if containerID, err := internal.GetContainerID(); err == nil {
+	if containerID, err := container.GetID(); err == nil {
 		defaultAttributes["container_id"] = containerID
 	}
