+++ sdk/internal/container/container.go
@@ -1,4 +1,4 @@
-package internal
+package container
 
 import (
 	"bufio"
@@ -13,7 +13,7 @@ const (
 	kubepodsPrefix = "/kubepods/"
 )
 
-// ErrNotInContainerEnv is returned when the GetContainerID function is
+// ErrNotInContainerEnv is returned when the GetID function is
 // called in a non container environment
 var ErrNotInContainerEnv = errors.New("not in a container environment")
 
@@ -35,8 +35,8 @@ func getContainerIDFromReader(f io.Reader) (string, error) {
 	return "", ErrNotInContainerEnv
 }
 
-// GetContainerID returns the container ID when in a containerized environment.
-func GetContainerID() (string, error) {
+// GetID returns the container ID when in a containerized environment.
+func GetID() (string, error) {
 	f, err := os.Open("/proc/self/cgroup")
 	if err != nil {
 		return "", err
