+++ sdk/internal/container/container_test.go
@@ -1,4 +1,4 @@
-package internal
+package container
 
 import (
 	"bytes"
