+++ tests/docker/Dockerfile
@@ -3,7 +3,7 @@ FROM golang:1.14
 WORKDIR $GOPATH/src/github.com/traceableai/goagent/docker
 
 COPY tests/docker/main.go main.go
-COPY sdk/internal/container.go internal/container.go
+COPY sdk/internal/container/container.go internal/container/container.go
 
 RUN go get -d -v ./...
